+++ components/automate-cli/pkg/verifyserver/server/api/v1/mockserver.go
@@ -38,7 +38,40 @@ func (h *Handler) StartMockServer(c *fiber.Ctx) error {
 		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
 	}
 
-	return c.Status(fiber.StatusOK).JSON(response.BuildSuccessResponse("Server started successfully"))
+	return c.JSON(response.BuildSuccessResponse("Server started successfully"))
+}
+
+func (h *Handler) StopMockServer(c *fiber.Ctx) error {
+
+	reqBody := new(models.StopMockServerRequestBody)
+	err := c.BodyParser(&reqBody)
+
+	// If request body is invalid
+	if err != nil {
+		h.Logger.Error("Stop mock server request body parsing failed: ", err.Error())
+		return fiber.NewError(fiber.StatusBadRequest, "Invalid request body")
+	}
+
+	// If port number is invalid
+	if !isValidPort(reqBody.Port) {
+		h.Logger.Error("Start mock-server request body contains invalid port number")
+		return fiber.NewError(fiber.StatusBadRequest, "Invalid port number")
+	}
+
+	err = h.MockServersService.Stop(reqBody)
+
+	if err != nil {
+		if strings.Contains(err.Error(), "no mock server found") {
+			// No Server is running on given port.
+			resp := fmt.Sprintf("No Mock server is running on port %d with protocol %v", reqBody.Port, reqBody.Protocol)
+			h.Logger.Debug(resp)
+			return fiber.NewError(fiber.StatusNotFound, resp)
+		}
+		h.Logger.Error("Error while stoppping server: ", err.Error())
+		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
+	}
+	return c.JSON(response.BuildSuccessResponse("Server stopped successfully"))
+
 }
 
 func isValidPort(requestBodyPort int) bool {
