+++ components/automate-cli/pkg/verifyserver/server/api/v1/mockserver_test.go
@@ -18,8 +18,17 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func SetupMockServerServiceForStop(protocol string) *mockserverservice.MockServersServiceMock {
+func SetupMockServerService(protocol string) *mockserverservice.MockServersServiceMock {
 	return &mockserverservice.MockServersServiceMock{
+		StartFunc: func(cfg *models.StartMockServerRequestBody) error {
+			if protocol == constants.UDP || protocol == constants.HTTPS {
+				return nil
+			}
+			if protocol == constants.TCP || protocol == constants.HTTP {
+				return errors.New("port unavailable")
+			}
+			return errors.New("unsupported protocol")
+		},
 		StopFunc: func(cfg *models.StopMockServerRequestBody) error {
 			if protocol == constants.TCP {
 				return errors.New("Error while shutting down the server")
@@ -32,7 +41,7 @@ func SetupMockServerServiceForStop(protocol string) *mockserverservice.MockServe
 	}
 }
 
-func SetupStopMockServerHandlers(ss mockserverservice.MockServersService) (*fiber.App, error) {
+func SetupMockServerHandlers(ss *mockserverservice.MockServersServiceMock) (*fiber.App, error) {
 	log, err := logger.NewLogger("text", "debug")
 	if err != nil {
 		return nil, err
@@ -44,7 +53,6 @@ func SetupStopMockServerHandlers(ss mockserverservice.MockServersService) (*fibe
 	app := fiber.New(fconf)
 	handler := v1.NewHandler(log).
 		AddMockServerService(ss)
-
 	vs := &server.VerifyServer{
 		Port:    server.DEFAULT_PORT,
 		Log:     log,
@@ -53,6 +61,97 @@ func SetupStopMockServerHandlers(ss mockserverservice.MockServersService) (*fibe
 	}
 	vs.SetupRoutes()
 	return vs.App, nil
+}
+
+func TestStartMockServer(t *testing.T) {
+	tests := []struct {
+		description  string
+		expectedCode int
+		expectedBody string
+		reqBody      string
+		protocol     string
+	}{
+		{
+			description:  "200:success status route",
+			expectedCode: 200,
+			protocol:     "udp",
+			reqBody: `{
+				"port": 3001,
+				"protocol": "udp",
+				"key": "",
+				"cert": ""
+		}`,
+		},
+		{
+			description:  "400:invalid request body",
+			expectedCode: 400,
+			protocol:     "udp",
+			reqBody: `{
+				"port": true,
+				"proto": "udp",
+				"key": "",
+				"cert": ""
+		}`,
+		},
+		{
+			description:  "409:port already in use",
+			expectedCode: 409,
+			protocol:     "tcp",
+			reqBody: `{
+				"port": 3000,
+				"protocol": "tcp",
+				"key": "",
+				"cert": ""
+		}`,
+		},
+		{
+			description:  "400:port invalid",
+			expectedCode: 400,
+			protocol:     "http",
+			reqBody: `{
+				"port": 65536,
+				"protocol": "http",
+				"key": "",
+				"cert": ""
+		}`,
+		},
+		{
+			description:  "400:port invalid",
+			expectedCode: 400,
+			protocol:     "udp",
+			reqBody: `{
+				"port": -1,
+				"key": "",
+				"cert": ""
+		}`,
+		},
+		{
+			description:  "500:internal server error",
+			expectedCode: 500,
+			protocol:     "dp",
+			reqBody: `{
+				"port": 4200,
+				"protocol": "dp",
+				"key": "",
+				"cert": ""
+		}`,
+		},
+	}
+	statusEndpoint := "/api/v1/start/mock-server"
+
+	for _, test := range tests {
+		app, err := SetupMockServerHandlers(SetupMockServerService(test.protocol))
+		assert.NoError(t, err)
+		reqBody := bytes.NewBufferString(test.reqBody)
+		req := httptest.NewRequest("POST", statusEndpoint, reqBody)
+		req.Header.Add("Content-Type", "application/json")
+		res, err := app.Test(req, -1)
+		assert.NoError(t, err)
+		body, err := io.ReadAll(res.Body)
+		assert.NoError(t, err, test.description)
+		assert.Contains(t, string(body), test.expectedBody)
+		assert.Equal(t, res.StatusCode, test.expectedCode)
+	}
 }
 
 func TestStopMockServerAPI(t *testing.T) {
@@ -76,7 +175,7 @@ func TestStopMockServerAPI(t *testing.T) {
 		{
 			description:  "400:bad request",
 			expectedCode: 400,
-			expectedBody: "{\"status\":\"FAILED\",\"result\":null,\"error\":{\"code\":400,\"message\":\"Invalid Body Request\"}}",
+			expectedBody: "{\"status\":\"FAILED\",\"result\":null,\"error\":{\"code\":400,\"message\":\"Invalid request body\"}}",
 			protocol:     constants.UDP,
 			reqBody: `{
 				"port": 5431,
@@ -119,7 +218,7 @@ func TestStopMockServerAPI(t *testing.T) {
 	for _, test := range tests {
 
 		// Setup the app as it is done in the main function
-		app, err := SetupStopMockServerHandlers(SetupMockServerServiceForStop(test.protocol))
+		app, err := SetupMockServerHandlers(SetupMockServerService(test.protocol))
 		assert.NoError(t, err)
 
 		reqBody := bytes.NewBufferString(test.reqBody)
